Bugfix: [export] Adding a filter on internal type articles (kind = 3)

+++ Components/LengowConfiguration.php
@@ -296,7 +296,7 @@ class Shopware_Plugins_Backend_Lengow_Components_LengowConfiguration
      */
     public static function getAccessIds()
     {
-        $accountId = self::getConfig('lengowAccountId');
+        $accountId = (int)self::getConfig('lengowAccountId');
         $accessToken = self::getConfig('lengowAccessToken');
         $secretToken = self::getConfig('lengowSecretToken');
         return array($accountId, $accessToken, $secretToken);
